use iso_fortran_env int64 for variables related to dict
This will increase the number of nact values that can be set, even if you build with 32-bit integers.

+++ src/density_matrix.f90
@@ -4,17 +4,18 @@ subroutine dim1_density(creat1, anhi1, sr, si)
 
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_dict
     use module_global_variables
 
     Implicit NONE
-    integer, intent(in) :: creat1, anhi1
+    integer(kind=int64), intent(in) :: creat1, anhi1
     real(8), intent(out) :: sr, si
 
-    integer :: j0, i, i0
+    integer(kind=int64) :: j0, i, i0
     complex*16 :: cmplxcii, cmplxcij, cmplxs
 
-    integer ::  newcas_idx, phase, phasenew
+    integer(kind=int64) ::  newcas_idx, phase, phasenew
 
 ! calculation of <0|Ec1a1|0>
 
@@ -53,21 +54,21 @@ subroutine dim1_density(creat1, anhi1, sr, si)
 
 end subroutine dim1_density
 
-
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 subroutine dim2_density(creat1, anhi1, creat2, anhi2, sr, si)
 
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_dict
     use module_global_variables
 
     Implicit NONE
-    integer, intent(in) :: creat2, anhi2, anhi1, creat1
+    integer(kind=int64), intent(in) :: creat2, anhi2, anhi1, creat1
     real(8), intent(out) :: sr, si
-    integer :: newcas_idx, phase, phasenew
-    integer :: j0, i, i0
+    integer(kind=int64) :: newcas_idx, phase, phasenew
+    integer(kind=int64) :: j0, i, i0
     complex*16 :: cmplxcii, cmplxcij, cmplxs
 
 ! calculation of <0|Ec1a1Ec2a2|0>
@@ -119,14 +120,15 @@ subroutine dim3_density(creat1, anhi1, creat2, anhi2, creat3, anhi3, sr, si)
 
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_dict
     use module_global_variables
 
     Implicit NONE
-    integer, intent(in) :: anhi3, anhi2, anhi1, creat3, creat2, creat1
+    integer(kind=int64), intent(in) :: anhi3, anhi2, anhi1, creat3, creat2, creat1
     real(8), intent(out) :: sr, si
-    integer :: newcas_idx, phase, phasenew
-    integer :: j0, i, i0
+    integer(kind=int64) :: newcas_idx, phase, phasenew
+    integer(kind=int64) :: j0, i, i0
     complex*16 :: cmplxcii, cmplxcij, cmplxs
 
 ! calculation of <0|Ec1a1Ec2a2Ec3a3|0>
@@ -181,14 +183,15 @@ subroutine dim4_density(creat1, anhi1, creat2, anhi2, creat3, anhi3, creat4, anh
 
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_dict
     use module_global_variables
 
     Implicit NONE
-    integer, intent(in) :: anhi4, anhi3, anhi2, anhi1, creat4, creat3, creat2, creat1
+    integer(kind=int64), intent(in) :: anhi4, anhi3, anhi2, anhi1, creat4, creat3, creat2, creat1
     real(8), intent(out) :: sr, si
-    integer :: newcas_idx, phase, phasenew
-    integer :: j0, i, i0
+    integer(kind=int64) :: newcas_idx, phase, phasenew
+    integer(kind=int64) :: j0, i, i0
     complex*16 :: cmplxcii, cmplxcij, cmplxs
 
 ! calculation of <0|Ec1a1Ec2a2Ec3a3Ec4a4|0>
@@ -249,15 +252,16 @@ subroutine dim1_density_R(creat1, anhi1, sr)
 
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_dict
     use module_global_variables
 
     Implicit NONE
-    integer, intent(in) :: creat1, anhi1
+    integer(kind=int64), intent(in) :: creat1, anhi1
     real(8), intent(out) :: sr
-    integer :: j0, i, i0
+    integer(kind=int64) :: j0, i, i0
 
-    integer ::  newcas_idx, phase, phasenew
+    integer(kind=int64) ::  newcas_idx, phase, phasenew
 
 ! calculation of <0|Ec1a1|0> for REAL
 
@@ -294,14 +298,15 @@ subroutine dim2_density_R(creat1, anhi1, creat2, anhi2, sr)
 
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_dict
     use module_global_variables
 
     Implicit NONE
-    integer, intent(in) :: creat2, anhi2, anhi1, creat1
+    integer(kind=int64), intent(in) :: creat2, anhi2, anhi1, creat1
     real(8), intent(out) :: sr
-    integer :: newcas_idx, phase, phasenew
-    integer :: j0, i, i0
+    integer(kind=int64) :: newcas_idx, phase, phasenew
+    integer(kind=int64) :: j0, i, i0
 
 ! calculation of <0|Ec1a1Ec2a2|0>
 
@@ -345,14 +350,15 @@ subroutine dim3_density_R(creat1, anhi1, creat2, anhi2, creat3, anhi3, sr)
 
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_dict
     use module_global_variables
 
     Implicit NONE
-    integer, intent(in) :: anhi3, anhi2, anhi1, creat3, creat2, creat1
+    integer(kind=int64), intent(in) :: anhi3, anhi2, anhi1, creat3, creat2, creat1
     real(8), intent(out) :: sr
-    integer :: newcas_idx, phase, phasenew
-    integer :: j0, i, i0
+    integer(kind=int64) :: newcas_idx, phase, phasenew
+    integer(kind=int64) :: j0, i, i0
 
 ! calculation of <0|Ec1a1Ec2a2Ec3a3|0>
 
@@ -399,14 +405,15 @@ subroutine dim4_density_R(creat1, anhi1, creat2, anhi2, creat3, anhi3, creat4, a
 
 ! ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+    use, intrinsic :: iso_fortran_env, only: int64
     use module_dict
     use module_global_variables
 
     Implicit NONE
-    integer, intent(in) :: anhi4, anhi3, anhi2, anhi1, creat4, creat3, creat2, creat1
+    integer(kind=int64), intent(in) :: anhi4, anhi3, anhi2, anhi1, creat4, creat3, creat2, creat1
     real(8), intent(out) :: sr
-    integer :: newcas_idx, phase, phasenew
-    integer :: j0, i, i0
+    integer(kind=int64) :: newcas_idx, phase, phasenew
+    integer(kind=int64) :: j0, i, i0
 
 ! calculation of <0|Ec1a1Ec2a2Ec3a3Ec4a4|0>
